Moved reusable modal sheets and content to new package which shows preview of actor/movie details.
