+++ app/src/main/java/com/developersbreach/composeactors/ui/modalSheet/SheetContentActorDetails.kt
@@ -1,0 +1,24 @@
+package com.developersbreach.composeactors.ui.modalSheet
+
+import androidx.compose.foundation.layout.Column
+import androidx.compose.foundation.layout.fillMaxWidth
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Modifier
+import com.developersbreach.composeactors.model.ActorDetail
+
+/**
+ * TODO
+ * Complete modal sheet content for showing actor details instead of navigating to
+ * ActorDetails screen directly.
+ */
+@Composable
+fun SheetContentActorDetails(
+    actor: ActorDetail?,
+    selectedActor: (Int) -> Unit
+) {
+    Column(
+        modifier = Modifier.fillMaxWidth()
+    ) {
+
+    }
+}
